oe.path: added 'recurse' argument to remove()
This makes it possible to specify whether the equivalent of 'rm -rf' or
only this of 'rm -f' is wanted.  Due to backward compatibility it
defaults to the recursive variant.

Acked-by: Chris Larson <chris_larson@mentor.com>

+++ lib/oe/path.py
@@ -43,13 +43,13 @@ def format_display(path, metadata):
     else:
         return rel
 
-def remove(path):
+def remove(path, recurse=True):
     """Equivalent to rm -f or rm -rf"""
     import os, errno, shutil
     try:
         os.unlink(path)
     except OSError, exc:
-        if exc.errno == errno.EISDIR:
+        if recurse and exc.errno == errno.EISDIR:
             shutil.rmtree(path)
         elif exc.errno != errno.ENOENT:
             raise
